+++ web/presentation.js
@@ -3,6 +3,7 @@ import {
     ADVANCED_NODE_IDS,
     setupAdvancedWidgetCompatibility,
 } from "./advanced_widgets.js";
+import { migrateLegacyWorkflowData } from "./workflow_compat.js";
 
 function isTarget(node) {
     return ADVANCED_NODE_IDS.has(node.constructor?.comfyClass);
@@ -10,6 +11,7 @@ function isTarget(node) {
 
 app.registerExtension({
     name: "llamacpp.AdvancedWidgets",
+    beforeConfigureGraph: migrateLegacyWorkflowData,
     nodeCreated(node) {
         if (isTarget(node)) setupAdvancedWidgetCompatibility(node, { fit: true });
     },